refactor: rename and move contexts and hooks
